implem controller et quelques modifs model mapper service

+++ src/main/java/com/example/bety/service/PersonService.java
@@ -6,6 +6,9 @@ import com.example.bety.model.service.Person;
 import com.example.bety.repository.PersonRepo;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -13,6 +16,7 @@ public class PersonService {
 
     PersonRepo personRepo;
     PersonMapper personMapper;
+
     PersonService(PersonRepo personRepo, PersonMapper personMapper) {
         this.personRepo = personRepo;
         this.personMapper = personMapper;
@@ -26,20 +30,32 @@ public class PersonService {
     public Person addPerson(Person p) {
         PersonModel personModel = personMapper.service2Bdd(p);
         personModel.setId(null);
-        PersonModel resultat = personRepo.save(personModel);
-        return personMapper.bdd2Service(resultat);
+        PersonModel result = personRepo.save(personModel);
+        return personMapper.bdd2Service(result);
     }
 
     public boolean deletePerson(Long id) {
-        boolean reponse = personRepo.existsById(id);
-        personRepo.deleteById(id);
-        return reponse;
+        if (personRepo.existsById(id)) {
+            personRepo.deleteById(id);
+            return true;
+        }
+
+        return false;
     }
 
     public Person updatePerson(Person p) {
         PersonModel personModel = personMapper.service2Bdd(p);
-        PersonModel resultat = personRepo.save(personModel);
-        return personMapper.bdd2Service(resultat);
+        PersonModel result = personRepo.save(personModel);
+        return personMapper.bdd2Service(result);
     }
 
+    public List<Person> showAllPersons() {
+
+        Iterable<PersonModel> iPersonModel = personRepo.findAll();
+        List<Person> listPerson = new ArrayList<>();
+        for (PersonModel p : iPersonModel) {
+            listPerson.add(personMapper.bdd2Service(p));
+        }
+        return listPerson;
+    }
 }
